ffspart, libflash: Fix stack size warnings
    libflash/file.c: In function 'file_erase':
    libflash/file.c:134:1: error: the frame size of 4128 bytes is larger than 1024 bytes [-Werror=frame-larger-than=]
     }
     ^

and

    ffspart.c: In function ‘main’:
    ffspart.c:529:1: error: the frame size of 4864 bytes is larger than 1024 bytes [-Werror=frame-larger-than=]
     }
     ^

In both cases, mark the local variables as static to avoid the stack.

The static approach is valid for file.c as the buffer is always filled
with `~0`. Given it's now going to be in .bss due to static we have to
still perform the memset(), but racing memset()s in this fashion won't
be harmful, just wasteful.

For ffspart.c's main(), there are bigger problems if that needs to be
re-entrant.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>