libflash/ipmi-hiomap: Improve BMC error state behaviour
impi_hiomap_handle_events()'s implementation contained an early-exit
that was only valid for handling ackable events, but the state provided
in the BMC's notification contains non-ackable state as well. Remove the
early exit to ensure we don't attempt to proceed with requests that we
know will fail which occur after an initial request has processed the
ackable events.

Include a test case to capture the desired behaviour.

Cc: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>
Signed-off-by: Andrew Jeffery <andrew@aj.id.au>